+++ social_core/backends/base.py
@@ -70,11 +70,13 @@ class BaseAuth(object):
            'strategy' not in kwargs or 'response' not in kwargs:
             return None
 
+
         self.strategy = self.strategy or kwargs.get('strategy')
         self.redirect_uri = self.redirect_uri or kwargs.get('redirect_uri')
         self.data = self.strategy.request_data()
         kwargs.setdefault('is_new', False)
         pipeline = self.strategy.get_pipeline(self)
+        args, kwargs = self.strategy.clean_authenticate_args(*args, **kwargs)
         return self.pipeline(pipeline, *args, **kwargs)
 
     def pipeline(self, pipeline, pipeline_index=0, *args, **kwargs):